Mengedit Method Put dan Post

+++ src/main/java/praktikum/praktikum_10/ProductServiceController.java
@@ -42,17 +42,30 @@ public class ProductServiceController {
    
    @RequestMapping(value = "/products/{id}", method = RequestMethod.PUT)
    public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product) { 
-      productRepo.remove(id);
-      product.setId(id);
-      productRepo.put(id, product);
-      return new ResponseEntity<>("Product is updated successsfully", HttpStatus.OK);
-   }
+     
+       if(!productRepo.containsKey(id)){
+            return new ResponseEntity<>("Product Not Found, Please check again", HttpStatus.NOT_FOUND);
+        }
+        else{
+            productRepo.remove(id);
+            product.setId(id);
+            productRepo.put(id, product);
+            return new  ResponseEntity<>("Product is updated Successfully",HttpStatus.OK);
+        }
+        
+    }
    
     @RequestMapping(value = "/products", method = RequestMethod.POST)
    public ResponseEntity<Object> createProduct(@RequestBody Product product) {
-      productRepo.put(product.getId(), product);
-      return new ResponseEntity<>("Product is created successfully", HttpStatus.CREATED);
-   }
+      
+       if(productRepo.containsKey(product.getId())){
+            return new ResponseEntity<>("ID Product Cannot be the Same, please check again", HttpStatus.OK);
+        }
+        else{
+            productRepo.put(product.getId(), product);
+            return new ResponseEntity<>("Product is created Successfully", HttpStatus.CREATED);
+        }
+    }
    
    @RequestMapping(value = "/products")
    public ResponseEntity<Object> getProduct() {
